+++ src/test/eta/Test.hs
@@ -1,5 +1,7 @@
 module Main(main) where
 
+import HelVM.HelCam.BrainFuck.TokensTest
+
 import Test.HUnit
 
 testExample :: Test
@@ -7,6 +9,7 @@ testExample = TestCase (assertEqual "test" "test" "test")
 
 testList :: Test
 testList = TestList [ TestLabel "testExample" testExample
+                    , TestLabel "testsOfTokens" testsOfTokens
                     ]
 
 main :: IO ()